requirements.txt
streamlit
torch
transformers
opencv-python
pillow
librosa
numpy